etl/requirements.txt
firebird-driver==1.9.0
python-dotenv==1.0.1
